requirements.txt
# ByteStructChat
pynacl
